+++ src/main/java/com/example/auction/config/security/OAuth2SuccessHandler.java
@@ -1,0 +1,33 @@
+package com.example.auction.config.security;
+
+import java.io.IOException;
+
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+
+import lombok.RequiredArgsConstructor;
+
+import org.springframework.security.core.Authentication;
+import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
+import org.springframework.stereotype.Component;
+
+import com.example.auction.domain.user.entity.CustomOAuth2User;
+
+/**
+ * 성공 후 프론트에 전달 (쿼리파라미터)
+ */
+@Component
+@RequiredArgsConstructor
+public class OAuth2SuccessHandler implements AuthenticationSuccessHandler {
+
+	@Override
+	public void onAuthenticationSuccess(HttpServletRequest request, HttpServletResponse response,
+		Authentication authentication)
+		throws IOException {
+		CustomOAuth2User oAuth2User = (CustomOAuth2User)authentication.getPrincipal();
+		String token = oAuth2User.getJwtToken();
+
+		response.sendRedirect("http://localhost:3000/oauth2/success?token=" + token);
+	}
+
+}
